test: move param_name to common module and add name member concept to template
